Eliminate the 128-byte magic blob
The magic blob is just some arbitrary data that the boot ROM calculates
a CRC over to compare with a static value. So as long as the data sent
to the boot ROM has the same CRC value as what the boot ROM is looking
for, the actual data that gets sent doesn't matter.

[0x2c, 0x08] appears to be the shortest sequence of bytes that generates
the correct CRC value, so we can eliminate the need for the 128-byte
blob by sending those two bytes in its place.